apps/backend/src/requirements.txt
alembic>=1.18.4
flagger>=3.1.0
flasgger>=0.9.7.1
flask>=3.1.3
flask-sqlalchemy>=3.1.1
psycopg2-binary>=2.9.11
pydantic>=2.12.5
python-dotenv>=1.2.2